Added an initial coding guidelines document.
It is based on the Olympus document. Some sections like styling,
templating and translating have been removed. They might either
be added back later once we have settled on how we want to solve
those problems, or they can be documented in a separate file.

I have also removed information on the directory structure. I believe
that should be documented in a separate document.