Updated information on the @author tag.
It should be used in the file header and in class doc blocks. Under
certain conditions it might also be used on a method level. The
requirement for being mentioned as an author has been set as "at
least more than 5 lines of code, excluding comments, brackets, etc."
This serves as a guideline and might not be treated quite as strict.